fix dtype incompatibility in modernbert adapter

+++ hf_adapters/hf_common.py
@@ -945,7 +945,7 @@ def build_prefill_mask_right_padded(
     return mask
 
 
-def add_sliding_window_band(mask, sliding_window, dtype=torch.float16):
+def add_sliding_window_band(mask, sliding_window):
     """Restrict an additive bidirectional mask to a local ``±sliding_window`` band.
 
     ModernBERT alternates global (full) attention layers with local
@@ -965,7 +965,7 @@ def add_sliding_window_band(mask, sliding_window, dtype=torch.float16):
     padded_len = mask.shape[-1]
     idx = torch.arange(padded_len)
     off_band = (idx[:, None] - idx[None, :]).abs() > sliding_window  # [L, L]
-    band = torch.zeros((padded_len, padded_len), dtype=dtype)
+    band = torch.zeros((padded_len, padded_len), dtype=mask.dtype)
     band[off_band] = -torch.inf
     return mask + band[None, None, :, :]
